Move planner prop tests and workflow helpers into separate files
